:bug: 修复当没有字典项时，无法正常删除字典的 bug

+++ ballcat-admin/ballcat-admin-core/src/main/java/com/hccake/ballcat/admin/modules/sys/service/impl/SysDictItemServiceImpl.java
@@ -49,7 +49,11 @@ public class SysDictItemServiceImpl extends ExtendServiceImpl<SysDictItemMapper,
 	 */
 	@Override
 	public boolean removeByDictCode(String dictCode) {
-		return baseMapper.deleteByDictCode(dictCode);
+		// 如果存在字典项则进行删除
+		if (baseMapper.existsDictItem(dictCode)) {
+			return baseMapper.deleteByDictCode(dictCode);
+		}
+		return true;
 	}
 
 }
